refactor: centralize CLI model access display
